# backend/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HTL Kahoot API",
        "description": "API documentation for the HTL Kahoot application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/questions/{questionId}/media": {
            "post": {
                "tags": [
                    "Question Media"
                ],
                "summary": "Upload media for a question",
                "description": "Uploads an image or video file and attaches it to the question.",
                "operationId": "321871b507257e4001c8b6369259e435",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "type"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "image",
                                            "video",
                                            "code_snippet"
                                        ]
                                    },
                                    "alt_text": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "sort_order": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Media uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/QuestionMedia"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/questions/{questionId}/media/{mediaId}": {
            "delete": {
                "tags": [
                    "Question Media"
                ],
                "summary": "Delete question media",
                "description": "Deletes a media file from the question and from storage.",
                "operationId": "17d6a8c5ef1d3253c0283fe28fe530b2",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "mediaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "QuestionMedia": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "question_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "image",
                            "video",
                            "code_snippet"
                        ]
                    },
                    "url": {
                        "type": "string"
                    },
                    "alt_text": {
                        "type": "string",
                        "nullable": true
                    },
                    "sort_order": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Sanctum token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}